Ah ha! I think I solved the table population.
The label was given the value of the first passed value:
in my case, I tried to make those empty strings.

This made all the ids blank as well.

Changed the blanks to the reasonable names of the fields and
hoping tha they table elements are still going to be missing.